requirements.txt
streamlit>=1.25
openai>=1.0.0
python-dotenv>=1.0.0
